Identify unk_02055808 -> field_transition, merge with unk_0205578C
