--- include/unk_0205578C.h
@@ -1,8 +1,0 @@
-#ifndef POKEPLATINUM_UNK_0205578C_H
-#define POKEPLATINUM_UNK_0205578C_H
-
-#include "field_task.h"
-
-void sub_020557DC(FieldTask *param0, int param1, int param2);
-
-#endif // POKEPLATINUM_UNK_0205578C_H
