+++ src/encounter.c
@@ -29,6 +29,7 @@
 #include "field_map_change.h"
 #include "field_overworld_state.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "game_records.h"
 #include "heap.h"
 #include "journal.h"
@@ -46,8 +47,6 @@
 #include "unk_0202F1D4.h"
 #include "unk_0203D1B8.h"
 #include "unk_020528D0.h"
-#include "unk_0205578C.h"
-#include "unk_02055808.h"
 #include "unk_020562F8.h"
 #include "unk_0206AFE0.h"
 #include "unk_0206CCB0.h"
