+++ src/meson.build
@@ -217,8 +217,7 @@ pokeplatinum_c = files(
     'unk_02054D00.c',
     'unk_020553DC.c',
     'unk_020556C4.c',
-    'unk_0205578C.c',
-    'unk_02055808.c',
+    'field_transition.c',
     'unk_020559DC.c',
     'unk_02055C50.c',
     'unk_020562F8.c',
