+++ src/scrcmd.c
@@ -102,6 +102,7 @@
 #include "field_script_context.h"
 #include "field_system.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "game_records.h"
 #include "heap.h"
 #include "inlines.h"
@@ -186,7 +187,6 @@
 #include "unk_020528D0.h"
 #include "unk_02052C6C.h"
 #include "unk_020553DC.h"
-#include "unk_02055808.h"
 #include "unk_020559DC.h"
 #include "unk_0205749C.h"
 #include "unk_0205B33C.h"
