+++ src/unk_0203D1B8.c
@@ -91,6 +91,7 @@
 #include "field_overworld_state.h"
 #include "field_system.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "game_options.h"
 #include "game_records.h"
 #include "heap.h"
@@ -122,7 +123,6 @@
 #include "unk_020366A0.h"
 #include "unk_02038FFC.h"
 #include "unk_020553DC.h"
-#include "unk_02055808.h"
 #include "unk_020559DC.h"
 #include "unk_0205B33C.h"
 #include "unk_0206AFE0.h"
