+++ src/unk_02050568.c
@@ -15,6 +15,7 @@
 #include "core_sys.h"
 #include "field_map_change.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "heap.h"
 #include "inlines.h"
 #include "location.h"
@@ -22,7 +23,6 @@
 #include "player_avatar.h"
 #include "system_flags.h"
 #include "unk_02005474.h"
-#include "unk_02055808.h"
 #include "unk_02056B30.h"
 #include "unk_02070428.h"
 #include "vars_flags.h"
