+++ src/unk_0206C0E8.c
@@ -15,13 +15,13 @@
 #include "core_sys.h"
 #include "field_map_change.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "heap.h"
 #include "location.h"
 #include "map_object.h"
 #include "player_avatar.h"
 #include "system_flags.h"
 #include "unk_02005474.h"
-#include "unk_02055808.h"
 #include "unk_02056B30.h"
 #include "unk_02070428.h"
 #include "vars_flags.h"
