+++ src/unk_020933F8.c
@@ -31,6 +31,7 @@
 #include "communication_information.h"
 #include "communication_system.h"
 #include "field_task.h"
+#include "field_transition.h"
 #include "game_records.h"
 #include "heap.h"
 #include "journal.h"
@@ -53,7 +54,6 @@
 #include "unk_0202CC64.h"
 #include "unk_0202F108.h"
 #include "unk_020363E8.h"
-#include "unk_02055808.h"
 #include "unk_0206CCB0.h"
 #include "unk_02094EDC.h"
 #include "unk_02095AF0.h"
